feat: add validator for scrapers
