+++ src/events/crawler/tasks/fetchSources/index.js
@@ -1,0 +1,6 @@
+import loadSources from './loadSources.js';
+import validateSources from './validateSources.js';
+
+const fetchSources = async args => loadSources(args).then(validateSources);
+
+export default fetchSources;
